+++ src/components/Comment/Reply/Reply/Reply.module.scss
@@ -32,7 +32,7 @@
       width: 1em;
       height: 1em;
       margin-left: 0.5em;
-      background-image: url('../../../assets/image/unlock.png');
+      background-image: url('../../../../assets/image/unlock.png');
       background-repeat: no-repeat;
       background-size: cover;
       cursor: pointer;
@@ -41,7 +41,7 @@
       width: 1em;
       height: 1em;
       margin-left: 0.5em;
-      background-image: url('../../../assets/image/lock.png');
+      background-image: url('../../../../assets/image/lock.png');
       background-repeat: no-repeat;
       background-size: cover;
       cursor: pointer;
