+++ src/components/Comment/Reply/Reply/Reply.tsx
@@ -1,7 +1,8 @@
 import React, { Fragment, useEffect, useRef, useState } from 'react';
+import { LoginProps } from '../../Comment';
 import css from './Reply.module.scss';
 
-const Reply = () => {
+const Reply: React.FC<LoginProps> = ({ loginId }) => {
   //textarea 처음에 비활성화 -> 수정 클릭 시 활성화
   const [isMyTextarea, setIsMyTextarea] = useState(true);
   const myTextarea = useRef<HTMLTextAreaElement>(null);
@@ -20,15 +21,15 @@ const Reply = () => {
     }
   };
   const [isLoginUser, setIsLoginUser] = useState(false);
-  const loginId: string | null = localStorage.getItem('id');
-  const replyUserId = '1';
+
+  const replyUserId = '2';
   useEffect(() => {
     if (loginId === replyUserId) {
       setIsLoginUser(true);
     } else {
       setIsLoginUser(false);
     }
-  }, []);
+  }, [loginId]);
 
   const handleModifyButton = () => {
     if (isLoginUser && isMyTextarea) {
@@ -46,28 +47,28 @@ const Reply = () => {
           <button className={css.newReply}>답글 달기</button>
         </Fragment>
       );
+    } else if (!isLoginUser) {
+      return <button className={css.newReply}>답글 달기</button>;
     } else if (isLoginUser && !isMyTextarea) {
       return <button className={css.setModify}>수정하기</button>;
     }
   };
 
   return (
-    <Fragment>
-      <div className={css.replyContainer}>
-        <div className={css.replyWriterInfo}>
-          <p className={css.replyWriterName}>작성자1</p>
-          <p className={css.replyDate}>2022년 12월 12일 오후 11:30</p>
-        </div>
-
-        <textarea
-          className={css.replyContent}
-          disabled={isMyTextarea}
-          rows={1}
-          defaultValue={'저는 공개되어있는 댓글입니다..'}
-        />
-        {handleModifyButton()}
+    <div className={css.replyContainer}>
+      <div className={css.replyWriterInfo}>
+        <p className={css.replyWriterName}>작성자1</p>
+        <p className={css.replyDate}>2022년 12월 12일 오후 11:30</p>
       </div>
-    </Fragment>
+
+      <textarea
+        className={css.replyContent}
+        disabled={isMyTextarea}
+        rows={1}
+        defaultValue="저는 공개되어있는 댓글입니다.."
+      />
+      {handleModifyButton()}
+    </div>
   );
 };
