+++ src/components/Header/Header.module.scss
@@ -4,7 +4,7 @@
   background-color: #fff;
   border-bottom: 1px solid #f4f4f4;
   display: flex;
-  justify-content: start;
+  justify-content: flex-start;
   padding-left: 2rem;
   & h1 {
     align-self: center;
